[ROCm] Converge the two W4A16 linear kernels onto the upstream shell
#1175 brought the fork's own W4A16 contribution back from upstream, leaving
two near-duplicate kernels: HybridW4A16LinearKernel (fork) and
RDNAHybridW4A16LinearKernel (upstream), each with its own copy of the Triton
kernel and packing helpers.

They were not interchangeable. Upstream's Python hands the skinny GEMM
unpacked activation-dtype zero points [N, groups], but the fork's C++ moved to
packed int32 zp [N/8, groups] to cut 4x the DRAM traffic on a memory-bound
kernel. Each side is self-consistent; the cross pairing #1175 created is not,
so upstream's kernel aborted on every asymmetric decode:

  RuntimeError: Zero points must be int32 or uint32 (packed 8x uint4 along
                dim 0), got c10::Half

That left test_rdna_hybrid_w4a16.py at 12 failures and
test_w4a16_kernel_selection.py at 2 (it asserts a selection that would have
crashed AWQ decode, and only passed because the fork's kernel won the
priority list).

Keep upstream's module, class and op name as the single implementation and
port the fork's contract and tuning into it: packed zp, gfx1151 cache-cliff
padding of weight and metadata rows, the v_and_or_b32 packed fp16 dequant with
its dtype-aware tile table, the scale/zp carrier, and the optional dequantized
prefill cache behind --w4a16-prefill-dequant. Upstream's gfx12x tile branch is
kept and now actually reachable -- the fork checked on_gfx1x() first, which
covers gfx12 and made its own gfx12 branch dead code.

Two upstream pieces are deliberately dropped: the K*M <= 32768 LDS gate, which
guards a C++ constraint this fork's kernel no longer has (it streams the
overflow from global), and _GFX1X_PREFILL_OVERRIDES, which was tuned against
upstream's scalar dequant body and does not transfer to the packed one.

Tested on gfx1151 (AMD Radeon 8060S, ROCm 7.15):
  test_rdna_hybrid_w4a16.py          83 passed (was 71 passed, 12 failed)
  test_w4a16_kernel_selection.py      2 passed (was 0 passed, 2 failed)
  test_hybrid_w4a16_triton.py,
    test_rocm_compressed_tensors_w4a16.py
                                     43 passed
  test_hip_w4a16.py, test_moe_gemm_w4a16.py,
    test_hybrid_w4a16_moe.py        196 passed
  test_hybrid_w4a16_perf.py         no regressions vs the unchanged gfx1151
                                    goldens; 290 out-of-band results are all
                                    improvements
